Alpha_Vantage_Pipeline/functions/loading/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
google-cloud-storage
requests==2.28.1
google-cloud-bigquery
pandas
pyarrow
